Bee Updates, KubeJS Updates, Trinket Update.
Had to change lost trinkets config as it was crashing when I was messing with files.

Added a temporary framework and sorting structure for bees, cleaned every single json we have. Scrubbed them.

Textures added for bees.

This was an astronomical amount of work, I hope this framework pays off.  Hopefully you have this set to require review, but everything worked in a fresh instance of the pack on my end.